TVG-102 Blueprint for Guide endpoints
